requirements.txt
pytesseract
python-dotenv
Pillow
opencv-python
matplotlib
tzlocal
# ...add any other dependencies as needed...
